drama_agent/requirements.txt
openai>=1.0.0
httpx>=0.25.0
python-docx>=0.8.11
fastapi>=0.100.0
uvicorn>=0.22.0
